refactor PictureRepository
