--- app/src/main/java/com/github/HumanLearning2021/HumanLearningApp/offline/CachePictureRepository.kt
@@ -1,5 +1,0 @@
-package com.github.HumanLearning2021.HumanLearningApp.offline
-
-import android.content.Context
-
-class CachePictureRepository(dbName: String, context: Context): PictureRepository(dbName, context, context.cacheDir)
