+++ app/src/main/java/com/github/HumanLearning2021/HumanLearningApp/offline/CachePictureStorage.kt
@@ -1,0 +1,5 @@
+package com.github.HumanLearning2021.HumanLearningApp.offline
+
+import android.content.Context
+
+class CachePictureStorage(dbName: String, context: Context): PictureStorage(dbName, context, context.cacheDir)
